Update README.md
Introduction: 
Both Substitution cipher technique and Transposition cipher technique are the types of Traditional cipher which are used to convert the plain text into cipher text.

Substitution Cipher Technique:
In Substitution Cipher Technique plain text characters are replaced with other characters, numbers and symbols as well as in substitution Cipher Technique, character’s identity is changed while its position remains unchanged.

Transposition Cipher Technique:
Transposition Cipher Technique rearranges the position of the plain text’s characters. In transposition Cipher Technique, The position of the character is changed but character’s identity is not changed.
-----------------------------------------------------------------------------------------------------------------------------------------
Example : 
Caesar Cipher in Cryptography
Text : ABCDEFGHIJKLMNOPQRSTUVWXYZ
Shift: 23
Cipher: XYZABCDEFGHIJKLMNOPQRSTUVW

Text : ATTACKATONCE
Shift: 4
Cipher: EXXEGOEXSRGI
----------------------------------------------------------------------------------------------------------------------------------------
Algorithms:
Algorithm for Caesar Cipher:
Input:
A String of lower case letters, called Text.
An Integer between 0-25 denoting the required shift.
Procedure:
Traverse the given text one character at a time .
For each character, transform the given character as per the rule, depending on whether we’re encrypting or decrypting the text.
Return the new string generated.

Algorithm for Columnar Transposition Cipher:

Encryption
In a transposition cipher, the order of the alphabets is re-arranged to obtain the cipher-text.
The message is written out in rows of a fixed length, and then read out again column by column, and the columns are chosen in some scrambled order.
Width of the rows and the permutation of the columns are usually defined by a keyword.
For example, the word HACK is of length 4 (so the rows are of length 4), and the permutation is defined by the alphabetical order of the letters in the keyword. In this case, the order would be “3 1 2 4”.
Any spare spaces are filled with nulls or left blank or placed by a character (Example: _).
Finally, the message is read off in columns, in the order specified by the keyword.

Decryption
To decipher it, the recipient has to work out the column lengths by dividing the message length by the key length.
Then, write the message out in columns again, then re-order the columns by reforming the key word.